# code/rhserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/stock/insertRealTimeHq": {
            "post": {
                "description": "insert stock real-time data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "insertRealTimeHq",
                "parameters": [
                    {
                        "description": "stock real-time data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/insertStockHisHq": {
            "post": {
                "description": "insert stock history data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "insertStockHisHq",
                "parameters": [
                    {
                        "description": "stock history data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
...
        "/v1/stock/queryRealTimeHqByCode": {
            "get": {
                "description": "查询股票实时行情数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "queryRealTimeHqByCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query start time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query end time",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/queryStockHisHq": {
            "get": {
                "description": "查询股票实时行情数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "queryStockHisHq",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query start date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query end date",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query period, day, week, month",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "query adjust, qfq, hfq, empty for no adjust",
                        "name": "adjust",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "query order, order by createdate, 1 for asc, 0 for desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
...
        "/v1/stock/updateRealTimeHq": {
            "post": {
                "description": "update stock real-time data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "updateRealTimeHq",
                "parameters": [
                    {
                        "description": "stock real-time data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockRealTimeData"
                        }
                    }
                }
            }
        },
        "/v1/stock/updateStockHisHq": {
            "post": {
                "description": "update stock history data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock"
                ],
                "summary": "updateStockHisHq",
                "parameters": [
                    {
                        "description": "stock history data",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.StockHisHq"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mysql.StockHisHq": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "adjust": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "chgamount": {
                    "type": "number"
                },
                "close": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "low": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "pchg": {
                    "type": "number"
                },
                "period": {
                    "type": "string"
                },
                "turnover": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
...
        "mysql.StockRealTimeData": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "cap": {
                    "type": "number"
                },
                "chgamount": {
                    "type": "number"
                },
                "chginyear": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "createtime": {
                    "type": "string"
                },
                "fcap": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "low": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open": {
                    "type": "number"
                },
                "pchg": {
                    "type": "number"
                },
                "pe": {
                    "type": "number"
                },
                "preclose": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "turnover": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        }
    }
}